requirements.txt
streamlit
streamlit_authenticator
streamlit-aggrid
streamlit-keyup
df-global-search
pandas
plotly
pymongo
numpy
memoization
certifi
google-api-core
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
google-api-python-client
google-auth
google-auth-oauthlib